Week13/activity2/requirements.txt
google-genai>=1.46.0
pypdf>=4.2.0
python-docx>=1.1.0
